do not minify/uglify code
There's a strange bug where "minifying" the code actually makes it twice
the size. Unfortunately, `create-react-app` doesn't let you not minify
easily. Instead, I had to run `eject` and modify the Webpack config.
The reason this matters is that GitHub won't allow files larger than
100M, and the "minified" JavaScript was 140M.

Hopefully at some point in the future, this will be fixed and we can
un-eject the code.